requirements.txt
streamlit
langchain
langchain_chroma
langchain_community
langchain_groq
langchain_huggingface
tf-keras
pypdf
Chroma
chromadb
python-dotenv
transformers
pypdf
validators==0.28.1
youtube_transcript_api
unstructured
pytube
gradio